feat: Update vehicleId type to Long, add findByOrderId method, and clean up unused code

+++ BE_ChargeHub/src/main/java/swp391/code/swp391/dto/AvailableSlotsResponseDTO.java
@@ -36,7 +36,7 @@ public class AvailableSlotsResponseDTO {
     @AllArgsConstructor
     @Builder
     public static class VehicleInfo {
-        private Integer vehicleId;
+        private Long vehicleId;
         private String brand;
         private String model;
         private Double batteryCapacity; // kWh
